move options content into own folder
closes #328
