+++ projects/ngx-d3timeline/src/lib/activity-rectangle/selectors/activity-rectangle-position.selectors.ts
@@ -3,7 +3,7 @@ import {
   selectTimeScale,
   selectBandScale
 } from '../../store/timeline-selectors';
-import { selectTimeOrientation } from '../../options.selectors';
+import { selectTimeOrientation } from '../../options/options.selectors';
 import {
   getPositionInTimeAxis,
   getPositionInResourceAxis,
