+++ projects/ngx-d3timeline/src/lib/options/options-utils.ts
@@ -1,4 +1,4 @@
-import { Orientation } from './core/orientation';
+import { Orientation } from '../core/orientation';
 import { Options } from './options';
 
 export function getTimeOrientation(options: Options): Orientation {
