+++ projects/ngx-d3timeline/src/lib/options/options.ts
