+++ projects/ngx-d3timeline/src/lib/store/actions.ts
@@ -1,6 +1,6 @@
 import { Activity } from '../activity/activity';
 import { PositionedActivity } from '../activity/positioned-activity';
-import { Options } from '../options';
+import { Options } from '../options/options';
 import { identifier } from '../core/types';
 import { Action } from '../store-lib/action';
