+++ projects/ngx-d3timeline/src/lib/store/reducer.ts
@@ -4,7 +4,7 @@ import { View } from '../view/view';
 import { TimelineDragEvent } from '../timeline-drag-event';
 import { identifier } from '../core/types';
 import { initialisePositionedActivity } from '../activity/positioned-activity';
-import { defaultOptions, CompleteOptions } from '../options';
+import { defaultOptions, CompleteOptions } from '../options/options';
 
 export function reducer(state: State = initialState, action: Actions): State {
   switch (action.type) {
