+++ projects/ngx-d3timeline/src/public-api.ts
@@ -5,4 +5,4 @@
 export * from './lib/ngx-d3timeline.component';
 export * from './lib/ngx-d3timeline.module';
 export * from './lib/activity/activity';
-export * from './lib/options';
+export * from './lib/options/options';
